+++ R/COG_akinator.R
@@ -4,9 +4,9 @@
 #' @param vecteur_codgeo correspond à un vecteur de codes Insee communaux pour lequel on cherche à déterminer le millésime de la géographie.
 #' @param donnees_insee vaut TRUE si les données manipulées sont produites par l'Insee. En effet, quelques rares modifications communales (la défusion des communes Loisey et Culey au 1er janvier 2014 par exemple) ont été prises en compte dans les bases de données communales de l'Insee plus tard que la date officielle. Pour tous les COG officiels datant d'avant 2008, seules les tables de passage Insee sont disponibles dans ce package.
 #' @details
-#' Le code officiel géographique le plus récent du package est actuellement celui au 01/01/2020. \cr
+#' Le code officiel géographique le plus récent du package est actuellement celui au 01/01/2021. \cr
 #'
-#' Les millésimes des COG qui peuvent être utilisés sont à ce stade les suivants : 1968, 1975, 1982, 1990, 1999, 2008 à 2020. \cr
+#' Les millésimes des COG qui peuvent être utilisés sont à ce stade les suivants : 1968, 1975, 1982, 1990, 1999, 2008 à 2021. \cr
 #'
 #' Les dates de référence des codes officiels géographiques utilisés dans COGugaison sont les suivantes :
 #' \itemize{