+++ R/apparier_COG.R
@@ -39,8 +39,8 @@
 #' @examples
 #' ## Exemple 1
 #' COG_akinator(exemple_popcom$CODGEO,donnees_insee = TRUE)
-#' apparier_COG(vecteur_codgeo=c(exemple_popcom[which(exemple_popcom$CODGEO!="01001"),1],"XXXXX"), donnees_insee = T,COG=2014)
-#' appariement <- apparier_COG(vecteur_codgeo=c(exemple_popcom[which(exemple_popcom$CODGEO!="01001"),1],"XXXXX"), donnees_insee = T,COG=2014)
+#' apparier_COG(vecteur_codgeo=c(exemple_popcom[which(exemple_popcom$CODGEO!="01001"),1],"XXXXX"), donnees_insee = TRUE,COG=2014)
+#' appariement <- apparier_COG(vecteur_codgeo=c(exemple_popcom[which(exemple_popcom$CODGEO!="01001"),1],"XXXXX"), donnees_insee = TRUE,COG=2014)
 #' cat(appariement$absent_de_bdd)
 #' cat(appariement$absent_de_COG)
 #' # regarder le libellé des communes présentes dans le COG mais pas dans la base de données